added user specificity to notes endpoints

+++ routes/notes.js
@@ -13,10 +13,14 @@ router.use(passport.authenticate('jwt', {session: false, failWithError: true}));
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-  console.log(req.header);
+
   const { searchTerm, folderId, tagId } = req.query;
 
-  let filter = {};
+  const userId = req.user.id;
+  console.log(userId);
+  console.log(req.user);
+
+  let filter = {userId};
 
   if (searchTerm) {
     // filter.title = { $regex: searchTerm, $options: 'i' };
@@ -50,13 +54,15 @@ router.get('/', (req, res, next) => {
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
 
+  const userId = req.user.id;
+
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
     err.status = 400;
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findOne({_id: id, userId})
     .populate('tags')
     .then(result => {
       if (result) {
@@ -73,6 +79,9 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { title, content, folderId, tags = [] } = req.body;
+  const userId = req.user.id;
+  console.log(userId);
+  console.log(req.user);
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -97,10 +106,11 @@ router.post('/', (req, res, next) => {
     });
   }
 
-  const newNote = { title, content, folderId, tags };
+  const newNote = {userId, title, content, folderId, tags};
 
   Note.create(newNote)
     .then(result => {
+      console.log(userId);
       res.location(`${req.originalUrl}/${result.id}`)
         .status(201)
         .json(result);
@@ -114,6 +124,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags = [] } = req.body;
+  const {userId} = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -145,7 +156,7 @@ router.put('/:id', (req, res, next) => {
 
   const updateNote = { title, content, folderId, tags };
 
-  Note.findByIdAndUpdate(id, updateNote, { new: true })
+  Note.findByIdAndUpdate({_id: id, userId}, updateNote, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -161,6 +172,7 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
+  const{userId} = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -169,7 +181,7 @@ router.delete('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findByIdAndRemove(id)
+  Note.findByIdAndRemove({_id: id, userId})
     .then(() => {
       res.sendStatus(204);
     })
